splitting the node
